WIP - Use command-line program to create database.
